Refactor: move vote ability in acceptance specs to shared examples
