+++ spec/acceptance/questions/vote_spec.rb
@@ -6,59 +6,9 @@ feature 'Vote for question', %q{
   I want to be able to vote for question
 } do
 
-  given(:user)             { create(:user) }
-  given(:author)           { create(:user) }
-  given(:question)         { create(:question, user: author) }
+  given(:question)  { create(:question, user: author) }
+  given(:container) { '.question' }
+  given(:obj)       { 'Question' }
 
-  scenario 'Non-authenticated user tries to vote for question' do
-    visit question_path(question)
-
-    within '.question' do
-      expect(page).to_not have_content '+'
-    end
-  end
-
-  context 'Authenticated user' do
-    scenario 'try to vote for his own question', js: true do
-      sign_in(author)
-      visit question_path(question)
-
-      within '.question' do
-        expect(page).to_not have_content '+'
-      end
-    end
-
-    scenario 'try to vote for another user question positively', js: true do
-      sign_in(user)
-      visit question_path(question)
-
-      within '.question' do
-        click_on '+'
-        expect(page).to have_content 'Question rating: 1'
-      end
-    end
-
-    scenario 'try to vote for another user question negatively', js: true do
-      sign_in(user)
-      visit question_path(question)
-
-      within '.question' do
-        click_on '-'
-        expect(page).to have_content 'Question rating: -1'
-      end
-    end
-
-    scenario 'try to vote for another user question twice', js: true do
-      sign_in(user)
-      visit question_path(question)
-
-      within '.question' do
-        click_on '+'
-        wait_for_ajax
-        click_on '+'
-        wait_for_ajax
-        expect(page).to have_content 'Question rating: 0'
-      end
-    end
-  end
+  it_behaves_like 'Vote ability'
 end
